[build] Moving custom layer to mbgl/gl
