+++ include/mbgl/platform/gl_functions.hpp
@@ -14,8 +14,6 @@
 namespace mbgl {
 namespace platform {
 
-#ifdef MBGL_RENDER_BACKEND_OPENGL
-
 using GLbitfield = unsigned int;
 using GLboolean = unsigned char;
 using GLchar = char;
@@ -329,8 +327,6 @@ extern void (* const glPointSize)(GLfloat);
 /// Pointer to glRasterPos4d OpenGL function.
 extern void (* const glRasterPos4d)(GLdouble, GLdouble, GLdouble, GLdouble);
 #endif
-
-#endif // MBGL_RENDER_BACKEND_OPENGL
 
 #ifndef NDEBUG
 /// Check for GL errors and print on the console.
