+++ src/core-files.json
@@ -21,11 +21,15 @@
         "src/mbgl/gl/attribute.cpp",
         "src/mbgl/gl/command_encoder.cpp",
         "src/mbgl/gl/context.cpp",
+        "src/mbgl/gl/custom_layer.cpp",
+        "src/mbgl/gl/custom_layer_factory.cpp",
+        "src/mbgl/gl/custom_layer_impl.cpp",
         "src/mbgl/gl/debugging_extension.cpp",
         "src/mbgl/gl/enum.cpp",
         "src/mbgl/gl/index_buffer_resource.cpp",
         "src/mbgl/gl/object.cpp",
         "src/mbgl/gl/offscreen_texture.cpp",
+        "src/mbgl/gl/render_custom_layer.cpp",
         "src/mbgl/gl/render_pass.cpp",
         "src/mbgl/gl/renderer_backend.cpp",
         "src/mbgl/gl/texture.cpp",
@@ -37,7 +41,6 @@
         "src/mbgl/gl/vertex_buffer_resource.cpp",
         "src/mbgl/layermanager/background_layer_factory.cpp",
         "src/mbgl/layermanager/circle_layer_factory.cpp",
-        "src/mbgl/layermanager/custom_layer_factory.cpp",
         "src/mbgl/layermanager/fill_extrusion_layer_factory.cpp",
         "src/mbgl/layermanager/fill_layer_factory.cpp",
         "src/mbgl/layermanager/heatmap_layer_factory.cpp",
@@ -120,7 +123,6 @@
         "src/mbgl/renderer/image_manager.cpp",
         "src/mbgl/renderer/layers/render_background_layer.cpp",
         "src/mbgl/renderer/layers/render_circle_layer.cpp",
-        "src/mbgl/renderer/layers/render_custom_layer.cpp",
         "src/mbgl/renderer/layers/render_fill_extrusion_layer.cpp",
         "src/mbgl/renderer/layers/render_fill_layer.cpp",
         "src/mbgl/renderer/layers/render_heatmap_layer.cpp",
@@ -218,8 +220,6 @@
         "src/mbgl/style/layers/circle_layer.cpp",
         "src/mbgl/style/layers/circle_layer_impl.cpp",
         "src/mbgl/style/layers/circle_layer_properties.cpp",
-        "src/mbgl/style/layers/custom_layer.cpp",
-        "src/mbgl/style/layers/custom_layer_impl.cpp",
         "src/mbgl/style/layers/fill_extrusion_layer.cpp",
         "src/mbgl/style/layers/fill_extrusion_layer_impl.cpp",
         "src/mbgl/style/layers/fill_extrusion_layer_properties.cpp",
@@ -339,13 +339,14 @@
         "mbgl/gfx/renderable.hpp": "include/mbgl/gfx/renderable.hpp",
         "mbgl/gfx/renderer_backend.hpp": "include/mbgl/gfx/renderer_backend.hpp",
         "mbgl/gfx/rendering_stats.hpp": "include/mbgl/gfx/rendering_stats.hpp",
+        "mbgl/gl/custom_layer.hpp": "include/mbgl/gl/custom_layer.hpp",
+        "mbgl/gl/custom_layer_factory.hpp": "include/mbgl/gl/custom_layer_factory.hpp",
         "mbgl/gl/renderable_resource.hpp": "include/mbgl/gl/renderable_resource.hpp",
         "mbgl/gl/renderer_backend.hpp": "include/mbgl/gl/renderer_backend.hpp",
         "mbgl/i18n/collator.hpp": "include/mbgl/i18n/collator.hpp",
         "mbgl/i18n/number_format.hpp": "include/mbgl/i18n/number_format.hpp",
         "mbgl/layermanager/background_layer_factory.hpp": "include/mbgl/layermanager/background_layer_factory.hpp",
         "mbgl/layermanager/circle_layer_factory.hpp": "include/mbgl/layermanager/circle_layer_factory.hpp",
-        "mbgl/layermanager/custom_layer_factory.hpp": "include/mbgl/layermanager/custom_layer_factory.hpp",
         "mbgl/layermanager/fill_extrusion_layer_factory.hpp": "include/mbgl/layermanager/fill_extrusion_layer_factory.hpp",
         "mbgl/layermanager/fill_layer_factory.hpp": "include/mbgl/layermanager/fill_layer_factory.hpp",
         "mbgl/layermanager/heatmap_layer_factory.hpp": "include/mbgl/layermanager/heatmap_layer_factory.hpp",
@@ -443,7 +444,6 @@
         "mbgl/style/layer_properties.hpp": "include/mbgl/style/layer_properties.hpp",
         "mbgl/style/layers/background_layer.hpp": "include/mbgl/style/layers/background_layer.hpp",
         "mbgl/style/layers/circle_layer.hpp": "include/mbgl/style/layers/circle_layer.hpp",
-        "mbgl/style/layers/custom_layer.hpp": "include/mbgl/style/layers/custom_layer.hpp",
         "mbgl/style/layers/fill_extrusion_layer.hpp": "include/mbgl/style/layers/fill_extrusion_layer.hpp",
         "mbgl/style/layers/fill_layer.hpp": "include/mbgl/style/layers/fill_layer.hpp",
         "mbgl/style/layers/heatmap_layer.hpp": "include/mbgl/style/layers/heatmap_layer.hpp",
@@ -558,6 +558,7 @@
         "mbgl/gl/attribute.hpp": "src/mbgl/gl/attribute.hpp",
         "mbgl/gl/command_encoder.hpp": "src/mbgl/gl/command_encoder.hpp",
         "mbgl/gl/context.hpp": "src/mbgl/gl/context.hpp",
+        "mbgl/gl/custom_layer_impl.hpp": "src/mbgl/gl/custom_layer_impl.hpp",
         "mbgl/gl/debugging_extension.hpp": "src/mbgl/gl/debugging_extension.hpp",
         "mbgl/gl/defines.hpp": "src/mbgl/gl/defines.hpp",
         "mbgl/gl/draw_scope_resource.hpp": "src/mbgl/gl/draw_scope_resource.hpp",
@@ -568,6 +569,7 @@
         "mbgl/gl/object.hpp": "src/mbgl/gl/object.hpp",
         "mbgl/gl/offscreen_texture.hpp": "src/mbgl/gl/offscreen_texture.hpp",
         "mbgl/gl/program.hpp": "src/mbgl/gl/program.hpp",
+        "mbgl/gl/render_custom_layer.hpp": "src/mbgl/gl/render_custom_layer.hpp",
         "mbgl/gl/render_pass.hpp": "src/mbgl/gl/render_pass.hpp",
         "mbgl/gl/renderbuffer_resource.hpp": "src/mbgl/gl/renderbuffer_resource.hpp",
         "mbgl/gl/state.hpp": "src/mbgl/gl/state.hpp",
@@ -648,7 +650,6 @@
         "mbgl/renderer/image_manager_observer.hpp": "src/mbgl/renderer/image_manager_observer.hpp",
         "mbgl/renderer/layers/render_background_layer.hpp": "src/mbgl/renderer/layers/render_background_layer.hpp",
         "mbgl/renderer/layers/render_circle_layer.hpp": "src/mbgl/renderer/layers/render_circle_layer.hpp",
-        "mbgl/renderer/layers/render_custom_layer.hpp": "src/mbgl/renderer/layers/render_custom_layer.hpp",
         "mbgl/renderer/layers/render_fill_extrusion_layer.hpp": "src/mbgl/renderer/layers/render_fill_extrusion_layer.hpp",
         "mbgl/renderer/layers/render_fill_layer.hpp": "src/mbgl/renderer/layers/render_fill_layer.hpp",
         "mbgl/renderer/layers/render_heatmap_layer.hpp": "src/mbgl/renderer/layers/render_heatmap_layer.hpp",
@@ -710,7 +711,6 @@
         "mbgl/style/layers/background_layer_properties.hpp": "src/mbgl/style/layers/background_layer_properties.hpp",
         "mbgl/style/layers/circle_layer_impl.hpp": "src/mbgl/style/layers/circle_layer_impl.hpp",
         "mbgl/style/layers/circle_layer_properties.hpp": "src/mbgl/style/layers/circle_layer_properties.hpp",
-        "mbgl/style/layers/custom_layer_impl.hpp": "src/mbgl/style/layers/custom_layer_impl.hpp",
         "mbgl/style/layers/fill_extrusion_layer_impl.hpp": "src/mbgl/style/layers/fill_extrusion_layer_impl.hpp",
         "mbgl/style/layers/fill_extrusion_layer_properties.hpp": "src/mbgl/style/layers/fill_extrusion_layer_properties.hpp",
         "mbgl/style/layers/fill_layer_impl.hpp": "src/mbgl/style/layers/fill_layer_impl.hpp",
