+++ src/mbgl/gl/render_custom_layer.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
+#include <mbgl/gl/custom_layer_impl.hpp>
 #include <mbgl/renderer/render_layer.hpp>
-#include <mbgl/style/layers/custom_layer_impl.hpp>
 
 namespace mbgl {
