+++ src/mbgl/platform/gl_functions.cpp
@@ -7,15 +7,9 @@ namespace platform {
 
 #ifndef NDEBUG
 void glCheckError(const char* cmd, const char* file, int line) {
-#ifdef MBGL_RENDER_BACKEND_OPENGL
     if (GLenum err = glGetError()) {
         Log::Warning(Event::OpenGL, "Error %d: %s - %s:%d", err, cmd, file, line);
     }
-#else
-    (void)cmd;
-    (void)file;
-    (void)line;
-#endif
 }
 #endif
