+++ src/mbgl/tile/geometry_tile.cpp
@@ -1,25 +1,25 @@
 #include <mbgl/tile/geometry_tile.hpp>
 
-#include <mbgl/tile/geometry_tile_worker.hpp>
-#include <mbgl/tile/geometry_tile_data.hpp>
-#include <mbgl/tile/tile_observer.hpp>
-#include <mbgl/style/layer_impl.hpp>
-#include <mbgl/style/layers/background_layer.hpp>
-#include <mbgl/style/layers/custom_layer.hpp>
+#include <mbgl/actor/scheduler.hpp>
+#include <mbgl/geometry/feature_index.hpp>
+#include <mbgl/gl/custom_layer.hpp>
+#include <mbgl/gl/render_custom_layer.hpp>
+#include <mbgl/map/transform_state.hpp>
+#include <mbgl/renderer/buckets/symbol_bucket.hpp>
+#include <mbgl/renderer/image_atlas.hpp>
+#include <mbgl/renderer/layers/render_background_layer.hpp>
+#include <mbgl/renderer/layers/render_symbol_layer.hpp>
+#include <mbgl/renderer/query.hpp>
 #include <mbgl/renderer/render_source.hpp>
 #include <mbgl/renderer/tile_parameters.hpp>
-#include <mbgl/renderer/layers/render_background_layer.hpp>
-#include <mbgl/renderer/layers/render_custom_layer.hpp>
-#include <mbgl/renderer/layers/render_symbol_layer.hpp>
-#include <mbgl/renderer/buckets/symbol_bucket.hpp>
-#include <mbgl/renderer/query.hpp>
-#include <mbgl/text/glyph_atlas.hpp>
-#include <mbgl/renderer/image_atlas.hpp>
-#include <mbgl/geometry/feature_index.hpp>
-#include <mbgl/map/transform_state.hpp>
-#include <mbgl/util/logging.hpp>
-#include <mbgl/actor/scheduler.hpp>
 #include <mbgl/renderer/tile_render_data.hpp>
+#include <mbgl/style/layer_impl.hpp>
+#include <mbgl/style/layers/background_layer.hpp>
+#include <mbgl/text/glyph_atlas.hpp>
+#include <mbgl/tile/geometry_tile_data.hpp>
+#include <mbgl/tile/geometry_tile_worker.hpp>
+#include <mbgl/tile/tile_observer.hpp>
+#include <mbgl/util/logging.hpp>
 
 #include <mbgl/gfx/upload_pass.hpp>
