+++ test/api/custom_layer.test.cpp
@@ -1,14 +1,14 @@
 #include <mbgl/test/util.hpp>
 
-#include <mbgl/platform/gl_functions.hpp>
+#include <mbgl/gfx/headless_frontend.hpp>
+#include <mbgl/gl/custom_layer.hpp>
+#include <mbgl/gl/defines.hpp>
 #include <mbgl/map/map.hpp>
 #include <mbgl/map/map_options.hpp>
-#include <mbgl/gl/defines.hpp>
-#include <mbgl/gfx/headless_frontend.hpp>
+#include <mbgl/platform/gl_functions.hpp>
 #include <mbgl/storage/resource_options.hpp>
-#include <mbgl/style/style.hpp>
-#include <mbgl/style/layers/custom_layer.hpp>
 #include <mbgl/style/layers/fill_layer.hpp>
+#include <mbgl/style/style.hpp>
 #include <mbgl/util/io.hpp>
 #include <mbgl/util/mat4.hpp>
 #include <mbgl/util/run_loop.hpp>
